Replace std::vector usage to enable optimizations
The std::vector class requires the program to be built with C++ exceptions and RTTI enabled, which bloats it a lot.
Replace the usage with the Watcom WCValOrderedVector instead, although it does not provide a feature to reserve memory.
This allows removing some C++ features from the compilation flags, and returning the program to the Compact memory model.

Also add various other optimization flags, that (among other things) further reduce the size of the EXE file.